combine pic provider with content provider
