+++ src/Main.tsx
@@ -3,7 +3,6 @@ import { Provider } from 'react-redux';
 import { createRoot } from 'react-dom/client';
 import { GoogleOAuthProvider } from '@react-oauth/google';
 import { PersistGate } from 'redux-persist/integration/react';
-import { PictureProvider } from './providers/Pics.provider';
 import { App } from './App';
 import store from './redux/store';
 import { AuthProvider } from './providers/Auth.provider';
@@ -17,11 +16,9 @@ function Main() {
       <AuthProvider>
         <Provider store={store.store}>
           <PersistGate loading={null} persistor={store.persistor}>
-            <PictureProvider>
-              <ContentProvider>
-                <App />
-              </ContentProvider>
-            </PictureProvider>
+            <ContentProvider>
+              <App />
+            </ContentProvider>
           </PersistGate>
         </Provider>
       </AuthProvider>
